refactor: split files in commitstatus package per SCM provider
