+++ internal/commitstatus/commitstatus.go
@@ -1,18 +1,9 @@
 package commitstatus
 
 import (
-	"bytes"
 	"context"
-	"encoding/base64"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
-	"path"
 	"strings"
-	"time"
 )
 
 // State represents the commit status state sent to the Git provider.
@@ -145,551 +136,4 @@ func Get(ctx context.Context, provider Provider, repoURL, repoFullName, commitSH
 	default:
 		return getGitHubCompatible(ctx, baseURL, repoFullName, commitSHA, token, contextName)
 	}
-}
-
-// resolveProvider returns the effective provider to use, falling back to
-// URL-based detection when provider is ProviderAuto.
-func resolveProvider(provider Provider, host string) Provider {
-	if provider != ProviderAuto {
-		return provider
-	}
-
-	if isGitLab(host) {
-		return ProviderGitLab
-	}
-
-	if isGitHub(host) {
-		return ProviderGitHub
-	}
-
-	if isAzureDevOps(host) {
-		return ProviderAzureDevOps
-	}
-
-	return ProviderGitea
-}
-
-// isGitHub returns true for github.com (the SaaS product only; GitHub Enterprise
-// cannot be reliably detected by hostname and must be set explicitly).
-func isGitHub(host string) bool {
-	return bareHost(host) == "github.com"
-}
-
-// isGitLab returns true when the host is exactly gitlab.com.
-// Self-hosted GitLab at a custom domain requires GIT_SCM_PROVIDER=gitlab.
-// The host argument may include a port (e.g. "gitlab.com:443").
-func isGitLab(host string) bool {
-	return strings.ToLower(bareHost(host)) == "gitlab.com"
-}
-
-// isAzureDevOps returns true for Azure DevOps SaaS hosts.
-func isAzureDevOps(host string) bool {
-	h := strings.ToLower(bareHost(host))
-
-	return h == "dev.azure.com" || h == "ssh.dev.azure.com" || strings.HasSuffix(h, ".visualstudio.com")
-}
-
-// bareHost strips the port suffix from a host string (e.g. "example.com:8080" → "example.com").
-// IPv6 addresses in brackets are handled correctly.
-func bareHost(host string) string {
-	h := strings.ToLower(host)
-	if i := strings.LastIndex(h, ":"); i > 0 && strings.LastIndex(h, "]") < i {
-		return h[:i]
-	}
-
-	return h
-}
-
-// parseHostAndScheme extracts the hostname and scheme from a git repository URL.
-// Handles HTTPS, SSH (ssh://), and SCP-style git URLs (git@host:...).
-func parseHostAndScheme(rawURL string) (string, string, error) {
-	rawURL = strings.TrimSpace(rawURL)
-	if rawURL == "" {
-		return "", "", errors.New("empty URL")
-	}
-
-	// SCP-style SSH: git@host:owner/repo.git
-	if strings.Contains(rawURL, "@") && strings.Contains(rawURL, ":") && !strings.Contains(rawURL, "://") {
-		parts := strings.SplitN(rawURL, "@", 2)
-		if len(parts) == 2 {
-			hostPath := strings.SplitN(parts[1], ":", 2)
-			if len(hostPath) == 2 {
-				return hostPath[0], "https", nil
-			}
-		}
-	}
-
-	parsed, err := url.Parse(rawURL)
-	if err != nil {
-		return "", "", fmt.Errorf("invalid URL %q: %w", rawURL, err)
-	}
-
-	if parsed.Host == "" {
-		return "", "", fmt.Errorf("no host in URL %q", rawURL)
-	}
-
-	scheme := parsed.Scheme
-	if scheme == "" || scheme == "ssh" || scheme == "git" {
-		scheme = "https"
-	}
-
-	return parsed.Host, scheme, nil
-}
-
-// postGitHub posts a commit status using the GitHub REST API.
-// For github.com the public API (api.github.com) is used.
-// For any other host (GitHub Enterprise Server) the /api/v3 endpoint is used.
-func postGitHub(ctx context.Context, baseURL, host, repoFullName, commitSHA, token string, status Status) error {
-	var apiURL string
-
-	if bareHost(host) == "github.com" {
-		apiURL = fmt.Sprintf("https://api.github.com/repos/%s/statuses/%s", repoFullName, commitSHA)
-	} else {
-		// GitHub Enterprise Server: https://{hostname}/api/v3/repos/{owner}/{repo}/statuses/{sha}
-		apiURL = fmt.Sprintf("%s/api/v3/repos/%s/statuses/%s", baseURL, repoFullName, commitSHA)
-	}
-
-	type githubRequest struct {
-		State       string `json:"state"`
-		Description string `json:"description"`
-		Context     string `json:"context"`
-		TargetURL   string `json:"target_url,omitempty"`
-	}
-
-	body := githubRequest{
-		State:       string(status.State),
-		Description: status.Description,
-		Context:     status.Context,
-		TargetURL:   status.TargetURL,
-	}
-
-	return doPost(ctx, apiURL, bearerAuthToken(token), body)
-}
-
-func getGitHub(ctx context.Context, baseURL, host, repoFullName, commitSHA, token, contextName string) (Status, bool, error) {
-	var apiURL string
-
-	if bareHost(host) == "github.com" {
-		apiURL = fmt.Sprintf("https://api.github.com/repos/%s/statuses/%s", repoFullName, commitSHA)
-	} else {
-		apiURL = fmt.Sprintf("%s/api/v3/repos/%s/statuses/%s", baseURL, repoFullName, commitSHA)
-	}
-
-	return getGitHubStyle(ctx, apiURL, bearerAuthToken(token), contextName)
-}
-
-// postGitHubCompatible posts a commit status using the GitHub-compatible API.
-// This covers Gitea, Forgejo, and Gogs which share the same /api/v1 endpoint shape.
-func postGitHubCompatible(ctx context.Context, baseURL, _ /* host */, repoFullName, commitSHA, token string, status Status) error {
-	apiURL := fmt.Sprintf("%s/api/v1/repos/%s/statuses/%s", baseURL, repoFullName, commitSHA)
-
-	type githubRequest struct {
-		State       string `json:"state"`
-		Description string `json:"description"`
-		Context     string `json:"context"`
-		TargetURL   string `json:"target_url,omitempty"`
-	}
-
-	body := githubRequest{
-		State:       string(status.State),
-		Description: status.Description,
-		Context:     status.Context,
-		TargetURL:   status.TargetURL,
-	}
-
-	return doPost(ctx, apiURL, giteaAuthToken(token), body)
-}
-
-func getGitHubCompatible(ctx context.Context, baseURL, repoFullName, commitSHA, token, contextName string) (Status, bool, error) {
-	apiURL := fmt.Sprintf("%s/api/v1/repos/%s/statuses/%s", baseURL, repoFullName, commitSHA)
-	return getGitHubStyle(ctx, apiURL, giteaAuthToken(token), contextName)
-}
-
-func getGitHubStyle(ctx context.Context, apiURL, token, contextName string) (Status, bool, error) {
-	type githubStatus struct {
-		State       string `json:"state"`
-		Description string `json:"description"`
-		Context     string `json:"context"`
-		TargetURL   string `json:"target_url"`
-	}
-
-	var statuses []githubStatus
-	if err := doGet(ctx, apiURL, token, &statuses); err != nil {
-		return Status{}, false, err
-	}
-
-	for _, status := range statuses {
-		if status.Context != contextName {
-			continue
-		}
-
-		return Status{
-			State:       State(status.State),
-			Description: status.Description,
-			Context:     status.Context,
-			TargetURL:   status.TargetURL,
-		}, true, nil
-	}
-
-	return Status{}, false, nil
-}
-
-// postGitLab posts a commit status using the GitLab API.
-func postGitLab(ctx context.Context, baseURL, repoFullName, commitSHA, token string, status Status) error {
-	// GitLab requires the namespace/path to be URL-encoded with %2F separating path components.
-	encodedPath := strings.ReplaceAll(repoFullName, "/", "%2F")
-	apiURL := fmt.Sprintf("%s/api/v4/projects/%s/statuses/%s", baseURL, encodedPath, commitSHA)
-
-	// Map our states to GitLab pipeline states.
-	gitlabState := string(status.State)
-
-	switch status.State {
-	case StateFailure, StateError:
-		gitlabState = "failed"
-	case StatePending:
-		gitlabState = "running"
-	}
-
-	type gitlabRequest struct {
-		State       string `json:"state"`
-		Name        string `json:"name"`
-		Description string `json:"description"`
-		TargetURL   string `json:"target_url,omitempty"`
-	}
-
-	body := gitlabRequest{
-		State:       gitlabState,
-		Name:        status.Context,
-		Description: status.Description,
-		TargetURL:   status.TargetURL,
-	}
-
-	return doPost(ctx, apiURL, bearerAuthToken(token), body)
-}
-
-func getGitLab(ctx context.Context, baseURL, repoFullName, commitSHA, token, contextName string) (Status, bool, error) {
-	encodedPath := strings.ReplaceAll(repoFullName, "/", "%2F")
-	apiURL := fmt.Sprintf("%s/api/v4/projects/%s/repository/commits/%s/statuses", baseURL, encodedPath, commitSHA)
-
-	type gitlabStatus struct {
-		Status      string `json:"status"`
-		Name        string `json:"name"`
-		Description string `json:"description"`
-		TargetURL   string `json:"target_url"`
-	}
-
-	var statuses []gitlabStatus
-	if err := doGet(ctx, apiURL, bearerAuthToken(token), &statuses); err != nil {
-		return Status{}, false, err
-	}
-
-	for _, status := range statuses {
-		if status.Name != contextName {
-			continue
-		}
-
-		return Status{
-			State:       gitLabStateToCommitStatus(status.Status),
-			Description: status.Description,
-			Context:     status.Name,
-			TargetURL:   status.TargetURL,
-		}, true, nil
-	}
-
-	return Status{}, false, nil
-}
-
-func postAzureDevOps(ctx context.Context, baseURL, host, repoURL, repoFullName, commitSHA, token string, status Status) error {
-	apiURL, err := azureDevOpsStatusesURL(baseURL, host, repoURL, repoFullName, commitSHA)
-	if err != nil {
-		return err
-	}
-
-	type azureContext struct {
-		Name  string `json:"name"`
-		Genre string `json:"genre"`
-	}
-
-	type azureRequest struct {
-		State       string       `json:"state"`
-		Description string       `json:"description"`
-		Context     azureContext `json:"context"`
-		TargetURL   string       `json:"targetUrl,omitempty"`
-	}
-
-	body := azureRequest{
-		State:       commitStatusToAzureState(status.State),
-		Description: status.Description,
-		Context: azureContext{
-			Name:  status.Context,
-			Genre: "doco-cd",
-		},
-		TargetURL: status.TargetURL,
-	}
-
-	return doPost(ctx, apiURL, azureDevOpsAuthToken(token), body)
-}
-
-func getAzureDevOps(ctx context.Context, baseURL, host, repoURL, repoFullName, commitSHA, token, contextName string) (Status, bool, error) {
-	apiURL, err := azureDevOpsStatusesURL(baseURL, host, repoURL, repoFullName, commitSHA)
-	if err != nil {
-		return Status{}, false, err
-	}
-
-	type azureContext struct {
-		Name  string `json:"name"`
-		Genre string `json:"genre"`
-	}
-
-	type azureStatus struct {
-		State       string       `json:"state"`
-		Description string       `json:"description"`
-		TargetURL   string       `json:"targetUrl"`
-		Context     azureContext `json:"context"`
-	}
-
-	type azureListResponse struct {
-		Value []azureStatus `json:"value"`
-	}
-
-	var response azureListResponse
-	if err := doGet(ctx, apiURL, azureDevOpsAuthToken(token), &response); err != nil {
-		return Status{}, false, err
-	}
-
-	for _, status := range response.Value {
-		if status.Context.Name != contextName {
-			continue
-		}
-
-		return Status{
-			State:       azureStateToCommitStatus(status.State),
-			Description: status.Description,
-			Context:     status.Context.Name,
-			TargetURL:   status.TargetURL,
-		}, true, nil
-	}
-
-	return Status{}, false, nil
-}
-
-func azureDevOpsStatusesURL(baseURL, host, repoURL, repoFullName, commitSHA string) (string, error) {
-	projectPath, repository, err := parseAzureDevOpsProjectAndRepo(repoURL, repoFullName, host)
-	if err != nil {
-		return "", err
-	}
-
-	apiBase := strings.TrimSuffix(baseURL, "/")
-	if bareHost(host) == "ssh.dev.azure.com" {
-		apiBase = "https://dev.azure.com"
-	}
-
-	return fmt.Sprintf(
-		"%s/%s/_apis/git/repositories/%s/commits/%s/statuses?api-version=7.1",
-		apiBase,
-		projectPath,
-		url.PathEscape(repository),
-		commitSHA,
-	), nil
-}
-
-func parseAzureDevOpsProjectAndRepo(repoURL, repoFullName, host string) (string, string, error) {
-	segments := parseAzureDevOpsPathSegments(repoURL)
-
-	projectPath, repository := azureDevOpsProjectAndRepoFromSegments(segments)
-	if projectPath != "" && repository != "" {
-		return projectPath, repository, nil
-	}
-
-	fullNameSegments := splitPathSegments(repoFullName)
-
-	projectPath, repository = azureDevOpsProjectAndRepoFromSegments(fullNameSegments)
-	if projectPath != "" && repository != "" {
-		return projectPath, repository, nil
-	}
-
-	return "", "", fmt.Errorf("failed to parse Azure DevOps repository from %q (host=%q, full_name=%q)", repoURL, host, repoFullName)
-}
-
-func parseAzureDevOpsPathSegments(rawURL string) []string {
-	rawURL = strings.TrimSpace(rawURL)
-	if rawURL == "" {
-		return nil
-	}
-
-	if strings.Contains(rawURL, "@") && strings.Contains(rawURL, ":") && !strings.Contains(rawURL, "://") {
-		parts := strings.SplitN(rawURL, ":", 2)
-		if len(parts) != 2 {
-			return nil
-		}
-
-		return splitPathSegments(parts[1])
-	}
-
-	parsed, err := url.Parse(rawURL)
-	if err != nil {
-		return nil
-	}
-
-	return splitPathSegments(parsed.Path)
-}
-
-func azureDevOpsProjectAndRepoFromSegments(segments []string) (string, string) {
-	if len(segments) == 0 {
-		return "", ""
-	}
-
-	if segments[0] == "v3" && len(segments) >= 4 {
-		return path.Join(segments[1], segments[2]), strings.TrimSuffix(segments[3], ".git")
-	}
-
-	for i, segment := range segments {
-		if segment != "_git" || i == 0 || i+1 >= len(segments) {
-			continue
-		}
-
-		return path.Join(segments[:i]...), strings.TrimSuffix(segments[i+1], ".git")
-	}
-
-	return "", ""
-}
-
-func splitPathSegments(value string) []string {
-	value = strings.TrimSpace(value)
-
-	value = strings.Trim(value, "/")
-	if value == "" {
-		return nil
-	}
-
-	rawSegments := strings.Split(value, "/")
-	segments := make([]string, 0, len(rawSegments))
-
-	for _, segment := range rawSegments {
-		segment = strings.TrimSpace(segment)
-		if segment == "" {
-			continue
-		}
-
-		segments = append(segments, segment)
-	}
-
-	return segments
-}
-
-func gitLabStateToCommitStatus(state string) State {
-	switch strings.ToLower(strings.TrimSpace(state)) {
-	case "running", "pending":
-		return StatePending
-	case "success":
-		return StateSuccess
-	case "failed", "failure":
-		return StateFailure
-	default:
-		return StateError
-	}
-}
-
-func bearerAuthToken(token string) string {
-	return "Bearer " + token
-}
-
-func giteaAuthToken(token string) string {
-	return "token " + token
-}
-
-func azureDevOpsAuthToken(token string) string {
-	pat := ":" + strings.TrimSpace(token)
-
-	return "Basic " + base64.StdEncoding.EncodeToString([]byte(pat))
-}
-
-func commitStatusToAzureState(state State) string {
-	switch state {
-	case StatePending:
-		return "pending"
-	case StateSuccess:
-		return "succeeded"
-	case StateFailure:
-		return "failed"
-	default:
-		return "error"
-	}
-}
-
-func azureStateToCommitStatus(state string) State {
-	switch strings.ToLower(strings.TrimSpace(state)) {
-	case "pending", "inprogress":
-		return StatePending
-	case "succeeded", "success":
-		return StateSuccess
-	case "failed", "failure":
-		return StateFailure
-	default:
-		return StateError
-	}
-}
-
-func doPost(ctx context.Context, apiURL, authHeaderValue string, body any) error {
-	jsonData, err := json.Marshal(body)
-	if err != nil {
-		return fmt.Errorf("failed to marshal request: %w", err)
-	}
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, apiURL, bytes.NewBuffer(jsonData)) // #nosec G107
-	if err != nil {
-		return fmt.Errorf("failed to create request: %w", err)
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", authHeaderValue)
-
-	client := &http.Client{Timeout: 15 * time.Second}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to post commit status: %w", err)
-	}
-
-	defer func() {
-		_, _ = io.Copy(io.Discard, resp.Body)
-		_ = resp.Body.Close()
-	}()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return fmt.Errorf("commit status API returned %s for %s", resp.Status, apiURL)
-	}
-
-	return nil
-}
-
-func doGet(ctx context.Context, apiURL, authHeaderValue string, dst any) error {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, apiURL, nil) // #nosec G107
-	if err != nil {
-		return fmt.Errorf("failed to create request: %w", err)
-	}
-
-	req.Header.Set("Authorization", authHeaderValue)
-
-	client := &http.Client{Timeout: 15 * time.Second}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to get commit status: %w", err)
-	}
-
-	defer func() {
-		_, _ = io.Copy(io.Discard, resp.Body)
-		_ = resp.Body.Close()
-	}()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return fmt.Errorf("commit status API returned %s for %s", resp.Status, apiURL)
-	}
-
-	if err := json.NewDecoder(resp.Body).Decode(dst); err != nil {
-		return fmt.Errorf("failed to decode response: %w", err)
-	}
-
-	return nil
 }
